requirements.txt
fastapi
genai
pypdf
python-dotenv
python-multipart
uvicorn